ubfix

+++ src/ziplist.c
@@ -537,7 +537,7 @@ int zipTryEncoding(unsigned char *entry, unsigned int entrylen, long long *v, un
 /* Store integer 'value' at 'p', encoded as 'encoding' */
 void zipSaveInteger(unsigned char *p, int64_t value, unsigned char encoding) {
     int16_t i16;
-    int32_t i32;
+    uint32_t i32;
     int64_t i64;
     if (encoding == ZIP_INT_8B) {
         ((int8_t*)p)[0] = (int8_t)value;
@@ -546,7 +546,7 @@ void zipSaveInteger(unsigned char *p, int64_t value, unsigned char encoding) {
         memcpy(p,&i16,sizeof(i16));
         memrev16ifbe(p);
     } else if (encoding == ZIP_INT_24B) {
-        i32 = value<<8;
+        i32 = ((uint64_t) value)<<8;
         memrev32ifbe(&i32);
         memcpy(p,((uint8_t*)&i32)+1,sizeof(i32)-sizeof(uint8_t));
     } else if (encoding == ZIP_INT_32B) {
